requirements.txt
Flask==3.0.3
python-barcode==0.15.1
Pillow==10.4.0
